verbs: Implement ibv_query_gid and ibv_query_gid_type over ioctl
Currently ibv_query_gid and ibv_query_gid_type are implemented over
sysfs. In order to improve their performance we implement them using
the new query GID entry API, so now they will use ioctl and fallback
to sysfs.
